Move useEvent hook to file
